PR feedback: refactor io.helidon.webserver.netty into io.helidon.webserver
