+++ webserver/webserver/src/main/java/io/helidon/webserver/ByteBufRequestChunk.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
 import java.nio.ByteBuffer;
 import java.util.Objects;
