+++ webserver/webserver/src/main/java/io/helidon/webserver/UnboundedSemaphore.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
 import java.util.concurrent.atomic.AtomicLong;
