+++ webserver/webserver/src/test/java/io/helidon/webserver/OriginThreadPublisherTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
 /**
  * The OriginThreadPublisherTest.
