+++ webserver/webserver/src/test/java/io/helidon/webserver/PlainTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -25,10 +25,6 @@ import java.util.logging.Logger;
 
 import io.helidon.common.CollectionsHelper;
 import io.helidon.common.http.Http;
-import io.helidon.webserver.Handler;
-import io.helidon.webserver.Routing;
-import io.helidon.webserver.ServerConfiguration;
-import io.helidon.webserver.WebServer;
 import io.helidon.webserver.utils.SocketHttpClient;
 
 import org.hamcrest.collection.IsMapContaining;
