+++ webserver/webserver/src/test/java/io/helidon/webserver/SslTest.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.LogManager;
 import java.util.logging.Logger;
 
 import javax.net.ssl.SSLContext;
@@ -35,10 +34,6 @@ import javax.ws.rs.core.Response;
 import io.helidon.common.Builder;
 import io.helidon.common.configurable.Resource;
 import io.helidon.common.pki.KeyConfig;
-import io.helidon.webserver.Routing;
-import io.helidon.webserver.SSLContextBuilder;
-import io.helidon.webserver.ServerConfiguration;
-import io.helidon.webserver.WebServer;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
