+++ webserver/webserver/src/test/java/io/helidon/webserver/UnboundedSemaphoreTest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.netty;
+package io.helidon.webserver;
 
+import io.helidon.webserver.UnboundedSemaphore;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
